Add PocketBaseUML
Hey @benallfree, just wanted to contribute my repo to the list: [PocketBaseUML](https://pocketbase-uml.github.io/) is a free, open-source UML diagram generator for PocketBase. Though PocketBase already comes with a fantastic UI (better than Supabase or Firebase, IMO), I was missing the ability to actually SEE the data architecture in a graphical form, easy to parse and understand in a quick glance.